refactor(udp-protocol): move connect types to connect module
