+++ packages/udp-protocol/src/lib.rs
@@ -19,11 +19,13 @@
 #![allow(clippy::wildcard_imports)]
 
 pub mod common;
+pub mod connect;
 mod peer_id;
 pub mod request;
 pub mod response;
 
 pub use self::common::*;
+pub use self::connect::*;
 pub use self::peer_id::{PeerClient, PeerId};
 pub use self::request::*;
 pub use self::response::*;
